test(e2e): add steps
