+++ feature-runner/steps/device/registry.ts
@@ -1,33 +1,27 @@
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb'
 import {
-	codeBlockOrThrow,
 	regExpMatchedStep,
 	type StepRunner,
 } from '@nordicsemiconductor/bdd-markdown'
 import { Type } from '@sinclair/typebox'
-import mqtt from 'mqtt'
 import { randomUUID } from 'node:crypto'
-import { readFileSync } from 'node:fs'
-import path from 'node:path'
 import pRetry from 'p-retry'
 import { generateCode } from '@hello.nrfcloud.com/proto/fingerprint'
-import { getDevice as getDeviceFromIndex } from '../../devices/getDevice.js'
-import { getAttributesForDevice } from '../../devices/getAttributesForDevice.js'
-import { registerDevice } from '../../devices/registerDevice.js'
-import { registerUnsupportedDevice } from '../../devices/registerUnsupportedDevice.js'
-import type { getAllAccountsSettings } from '@hello.nrfcloud.com/nrfcloud-api-helpers/settings'
+import { getDevice as getDeviceFromIndex } from '../../../devices/getDevice.js'
+import { getAttributesForDevice } from '../../../devices/getAttributesForDevice.js'
+import { registerDevice } from '../../../devices/registerDevice.js'
+import { registerUnsupportedDevice } from '../../../devices/registerUnsupportedDevice.js'
 import { IMEI } from '@hello.nrfcloud.com/bdd-markdown-steps/random'
 
-const createDeviceForModel = ({
+export const createDeviceForModel = ({
 	db,
 	devicesTable,
 	devicesTableFingerprintIndexName,
 }: {
 	db: DynamoDBClient
-
 	devicesTable: string
 	devicesTableFingerprintIndexName: string
-}) =>
+}): StepRunner<Record<string, string>> =>
 	regExpMatchedStep(
 		{
 			regExp:
@@ -72,7 +66,6 @@ const createDeviceForModel = ({
 			progress(`Device registered: ${fingerprint} (${id})`)
 		},
 	)
-
 const waitForDeviceToBeAvailable =
 	({
 		db,
@@ -101,7 +94,6 @@ const waitForDeviceToBeAvailable =
 			},
 		)
 	}
-
 const waitForDeviceAttributesToBeAvailable =
 	({ db, devicesTable }: { db: DynamoDBClient; devicesTable: string }) =>
 	async (id: string) => {
@@ -121,8 +113,7 @@ const waitForDeviceAttributesToBeAvailable =
 			},
 		)
 	}
-
-const createUnsupportedDevice = ({
+export const createUnsupportedDevice = ({
 	db,
 	devicesTable,
 	devicesTableFingerprintIndexName,
@@ -131,7 +122,7 @@ const createUnsupportedDevice = ({
 
 	devicesTable: string
 	devicesTableFingerprintIndexName: string
-}) =>
+}): StepRunner<Record<string, string>> =>
 	regExpMatchedStep(
 		{
 			regExp:
@@ -163,64 +154,7 @@ const createUnsupportedDevice = ({
 		},
 	)
 
-const publishDeviceMessage = (
-	allAccountSettings: Awaited<ReturnType<typeof getAllAccountsSettings>>,
-) =>
-	regExpMatchedStep(
-		{
-			regExp:
-				/^the device `(?<id>[^`]+)` publishes this message to the MQTT topic `(?<topic>[^`]+)`$/,
-			schema: Type.Object({
-				id: Type.String(),
-				topic: Type.String(),
-			}),
-		},
-		async ({ match: { id, topic }, log: { progress, error }, step }) => {
-			const message = JSON.parse(codeBlockOrThrow(step).code)
-
-			const nRFCloudSettings = allAccountSettings['nordic']
-			if (nRFCloudSettings === undefined) {
-				throw new Error('No default nRF Cloud settings (nordic)')
-			}
-
-			progress(`Device id ${id} publishes to topic ${topic}`)
-			await new Promise((resolve, reject) => {
-				const mqttClient = mqtt.connect({
-					host: nRFCloudSettings.mqttEndpoint,
-					port: 8883,
-					protocol: 'mqtts',
-					protocolVersion: 4,
-					clean: true,
-					clientId: id,
-					key: nRFCloudSettings.accountDevicePrivateKey,
-					cert: nRFCloudSettings.accountDeviceClientCert,
-					ca: readFileSync(
-						path.join(process.cwd(), 'data', 'AmazonRootCA1.pem'),
-						'utf-8',
-					),
-				})
-
-				mqttClient.on('connect', () => {
-					progress('connected')
-					const mqttTopic = `${nRFCloudSettings.mqttTopicPrefix}${topic}`
-					progress('publishing', message, mqttTopic)
-					mqttClient.publish(mqttTopic, JSON.stringify(message), (error) => {
-						if (error) return reject(error)
-						mqttClient.end()
-						return resolve(void 0)
-					})
-				})
-
-				mqttClient.on('error', (err) => {
-					error(err)
-					reject(err)
-				})
-			})
-		},
-	)
-
 export const steps = (
-	allAccountSettings: Awaited<ReturnType<typeof getAllAccountsSettings>>,
 	db: DynamoDBClient,
 	{
 		devicesTableFingerprintIndexName,
@@ -233,5 +167,4 @@ export const steps = (
 		devicesTableFingerprintIndexName,
 		devicesTable,
 	}),
-	publishDeviceMessage(allAccountSettings),
 ]
