+++ package.json
@@ -65,6 +65,7 @@
     "@types/yazl": "2.4.5",
     "aws-cdk": "2.140.0",
     "aws-cdk-lib": "2.140.0",
+    "cbor-x": "1.5.9",
     "cdk": "2.140.0",
     "chalk": "5.3.0",
     "check-node-version": "4.2.1",
